rename: 로그인 API 응답 변경에 따른 DTO 디렉토리 변경
